Update Basic Results (For Caddy)
Earlier versions of the results were based on flawed configuration.
The flaw? Well, the `config` generated by the server was redirecting
directly to the server and not the proxy. In other words, even though
the initial connection was through the proxy, all the actual test
connections bypassed it.

The results in this commit are accurate.